Add files via upload
Here is the detailed Microsoft documentation for bulk installation of the new teams https://learn.microsoft.com/en-us/microsoftteams/new-teams-bulk-install-client

New teams is MSIX application and installation path is C:\Program Files\WindowsApps\MSTeams_24004.1304.2637.3318_x64__8wekyb3d8bbwe where as classic teams is located in %localAppdata%\Microsoft\Teams.

you can enable the new teams using Microsoft Teams update policy https://techcommunity.microsoft.com/t5/microsoft-teams-blog/enable-the-new-microsoft-teams-for-your-organization-today/ba-p/3945599

In anticipation of this transition, it is crucial to enable New Teams for users to explore its capabilities and provide feedback on any potential issues. 

I was tasked with identifying a method to monitor the number of users who have opted in for the installation of New Teams and have set Teams as their default Teams application.

In this blog post, we will discuss how to monitor the installation of New Teams and track whether users have set new Teams as their default instant messaging (IM) provider. This is useful till 31st march. after that, every user will have new teams as default and classic will not be functional.

Monitoring New Teams Installation:

Since the teams installation is MSIX application, it can be tracked using Get-AppxPackage *MSTeams* (this must run with user rights).

Monitoring Default IM Provider:

For Default IM provider, there is registry key in HKCU that helps to monitor if the default IM provider is new teams or classic.

For new Teams, the IMprovider value is MSTeams , where as for classic teams, it is Teams

we will use these values to monitor if the user has opt-in for new teams and is set as default Teams or not using intune remediation script feature.

Intune Remediation Script:

The Intune remediation script is available on GitHub repository. You can download, customize, and deploy it through Intune to start monitoring the New Teams installation and default IM provider settings.

Here is a brief overview of the key functionalities of the script:

1.Check Installation: The script verifies the presence of New Teams installation on users' devices.


2.Default IM Provider: It monitors registry settings to confirm whether users have set New Teams as their default IM provider.


3.Remediation: There is no remediation, it is only detection script. You can leave the remediation as blank.


Deploy the remediation to the group of devices with following settings (script run with user rights).